+++ src/seman/seman.rs
@@ -1,12 +1,13 @@
 use std::collections::{HashMap, HashSet};
 use std::fmt::Binary;
+use std::sync::mpsc::Sender;
 
 use crate::codegen::codegen::FuncData;
 use crate::fcparse::fcparse::AstNode;
 use crate::fcparse::fcparse::{self as fparse, AstRoot, BinaryOp, FuncArg, 
     FuncTable, UnaryOp};
 use crate::lex::Intrinsic;
-use crate::logger::logger as log;
+use crate::logger::logger::{self as log, LogMessage};
 use crate::logger::logger::{ErrKind, LogLevel, Logger, WarnKind};
 
 #[derive(Debug, Clone)]
@@ -161,7 +162,9 @@ pub type OverloadTable = HashMap<usize, (Option<usize>, FType)>;
 #[derive(Debug)]
 pub struct SemAn {
     symb_table: SymbolTable,
+    fname: String,
     cur_scope: usize,
+    module: String,
     permissive: bool,
     parsing_loop: Vec<usize>, // fold level
     func_table: FuncTable,
@@ -174,25 +177,32 @@ pub struct SemAn {
 
 impl SemAn {
     /// Inits semantic analyzer struct. Permissive flag for less type checks
-    pub fn new(permissive: bool, functab: FuncTable) -> SemAn {
+    pub fn new(permissive: bool, functab: FuncTable, fname: String) -> SemAn {
         SemAn {
             symb_table: SymbolTable::new(),
             cur_scope: 0,
             permissive: permissive,
             parsing_loop: Vec::new(),
             func_table: functab,
+            fname: fname,
             declared_parse: HashMap::new(),
             parsing_func: None,
             matched_overloads: HashMap::new(),
             expect_type: FType::none,
+            module: "main".to_owned()
         }
     }
 
-    pub fn analyze(&mut self, ast: &Vec<AstRoot>, logger: &mut Logger) {
+    pub fn analyze(&mut self, ast: &Vec<AstRoot>, logger: &Sender<LogMessage>) 
+        -> Result<(), ()> {
         match self.func_table.get_func("main") {
             Some(fv) => {
                 if fv.len() > 1 {
-                    logger.emit(LogLevel::Error(ErrKind::FewMains(fv.len())), 0);
+                    logger.send(LogMessage::new(
+                        LogLevel::Error(ErrKind::FewMains(fv.len())), 
+                        0,
+                        self.fname.clone()
+                    ));
                 }
             }
             None => {
@@ -203,9 +213,11 @@ impl SemAn {
         for root in ast {
             self.analyze_expr(root, logger);
         }
+
+        Ok(())
     }
 
-    fn analyze_expr(&mut self, node: &AstRoot, logger: &mut Logger) -> ExprDat {
+    fn analyze_expr(&mut self, node: &AstRoot, logger: &Sender<LogMessage>) -> ExprDat {
         let mut exprdat = ExprDat::new(FType::none);
         let line = node.line;
         match &node.node {
@@ -215,17 +227,19 @@ impl SemAn {
                 self.expect_type = FType::none;
 
                 if self.symb_table.get(name.clone()).is_some() {
-                    logger.emit(
+                    logger.send(LogMessage::new(
                         LogLevel::Error(ErrKind::Redeclaration(name.to_owned())),
                         line,
-                    );
+                        self.fname.clone()
+                    ));
                 };
 
                 if *ft == FType::none {
-                    logger.emit(
+                    logger.send(LogMessage::new(
                         LogLevel::Error(ErrKind::NoneTypeAssign(name.clone(), rightdat.ftype)),
                         line,
-                    );
+                        self.fname.clone()
+                    ));
                 }
 
                 match (*ft, rightdat.ftype) {
@@ -234,18 +248,20 @@ impl SemAn {
                     (FType::Array(fti1, c1, _), FType::Array(fti2, c2, _)) => {
                        if fti1 != fti2 {
                             println!("here");
-                           logger.emit(
+                           logger.send(LogMessage::new(
                                 LogLevel::Error(ErrKind::MismatchedTypes(*ft, rightdat.ftype)),
-                                    line,
-                           );
+                                line,
+                                self.fname.clone()
+                           ));
                        }
                     }
                     (other1, other2) => {
                         if other1 != other2 {
-                            logger.emit(
+                            logger.send(LogMessage::new(
                                 LogLevel::Error(ErrKind::MismatchedTypes(*ft, rightdat.ftype)),
                                 line,
-                            );
+                                self.fname.clone()
+                            ));
                         }
                     }
                 }
@@ -285,19 +301,21 @@ impl SemAn {
                         matches!(rightd.ftype, FType::uint | FType::heapptr);
 
                     if !is_valid_shift_right_type {
-                        logger.emit(
+                        logger.send(LogMessage::new(
                             LogLevel::Error(ErrKind::BitShiftRHSType(*op, rightd.ftype)),
                             line,
-                        );
+                            self.fname.clone()
+                        ));
                     }
                     // Note: left operand type doesn't need to match right operand for shifts
                 } else {
                     // For non-shift operations, types must match
                     if leftd.ftype != rightd.ftype {
-                        logger.emit(
+                        logger.send(LogMessage::new(
                             LogLevel::Error(ErrKind::BinOpDifTypes(*op, leftd.ftype, rightd.ftype)),
                             line,
-                        );
+                            self.fname.clone()
+                        ));
                     }
                 }
 
@@ -307,7 +325,11 @@ impl SemAn {
                         || *op == BinaryOp::Multiply
                         || *op == BinaryOp::Divide)
                 {
-                    logger.emit(LogLevel::Error(ErrKind::BoolBounds(*op)), line);
+                    logger.send(LogMessage::new(
+                        LogLevel::Error(ErrKind::BoolBounds(*op)), 
+                        line,
+                        self.fname.clone()
+                    ));
                 }
 
                 let is_cmp_op = SemAn::is_cmp_op(op);
@@ -324,17 +346,22 @@ impl SemAn {
                 if (*op == UnaryOp::Negate)
                     && !((exprdat.ftype == FType::double) || (exprdat.ftype == FType::int))
                 {
-                    logger.emit(LogLevel::Error(ErrKind::NegateBounds(exprdat.ftype)), line);
+                    logger.send(LogMessage::new(
+                        LogLevel::Error(ErrKind::NegateBounds(exprdat.ftype)),
+                        line,
+                        self.fname.clone()
+                    ));
                 }
             }
             AstNode::Variable(var) => {
                 exprdat.ftype = match self.symb_table.get(var.clone()) {
                     Some(v) => v.1.ftype,
                     None => {
-                        logger.emit(
+                        logger.send(LogMessage::new(
                             LogLevel::Error(ErrKind::UndeclaredVar(var.to_owned())),
                             line,
-                        );
+                            self.fname.clone()
+                        ));
                         FType::none
                     }
                 }
@@ -344,8 +371,11 @@ impl SemAn {
                     let entry = match self.symb_table.get(name.clone()) {
                         Some(en) => en,
                         None => {
-                            logger
-                                .emit(LogLevel::Error(ErrKind::UndeclaredVar(name.clone())), line);
+                            logger.send(LogMessage::new(
+                                LogLevel::Error(ErrKind::UndeclaredVar(name.clone())), 
+                                line,
+                                self.fname.clone()
+                            ));
                             (
                                 0,
                                 &FSymbol::new(name.to_owned(), VarPosition::None, FType::none),
@@ -362,10 +392,11 @@ impl SemAn {
                     other => other
                 };
                 if res_type != newval_data.ftype {
-                    logger.emit(
+                    logger.send(LogMessage::new(
                         LogLevel::Error(ErrKind::IncompatTypes(res_type, newval_data.ftype)),
                         line,
-                    );
+                        self.fname.clone()
+                    ));
                 }
             }
             AstNode::IfStatement {
@@ -377,7 +408,7 @@ impl SemAn {
                 if cond_an.ftype != FType::bool {
                     let lerr = LogLevel::Error(ErrKind::IfStmtNotBool(cond_an.ftype));
                     let lwarn = LogLevel::Warning(log::WarnKind::IfStmtNotBool);
-                    self.permissive_error(line, logger, lerr, lwarn);
+                    self.permissive_error(line, logger.clone(), lerr, lwarn);
                 }
 
                 self.analyze_expr(*&if_true, logger);
@@ -389,13 +420,18 @@ impl SemAn {
                 // TODO: add some meaningful type checks here
                 if let Some(var) = self.symb_table.get(name.clone()) {
                     if var.1.ftype == *target_type {
-                        logger.emit(LogLevel::Warning(WarnKind::ConvSame(var.1.ftype)), line);
+                        logger.send(LogMessage::new(
+                            LogLevel::Warning(WarnKind::ConvSame(var.1.ftype)), 
+                            line,
+                            self.fname.clone()
+                        ));
                     }
                 } else {
-                    logger.emit(
+                    logger.send(LogMessage::new(
                         LogLevel::Error(ErrKind::UndeclaredVar(name.to_owned())),
                         line,
-                    );
+                        self.fname.clone()
+                    ));
                 }
 
                 exprdat.ftype = *target_type;
@@ -408,12 +444,13 @@ impl SemAn {
                             FType::Array(_, _, _) => {},
                             FType::strconst => {},
                             other => {
-                                logger.emit(
+                                logger.send(LogMessage::new(
                                     LogLevel::Error(
                                         ErrKind::MismatchedTypes(FType::Array(0, 0, 0), other)
                                     ),
-                                    line
-                                );
+                                    line,
+                                    self.fname.clone()
+                                ));
                             }
                         }
                         exprdat.ftype = FType::uint; 
@@ -435,7 +472,7 @@ impl SemAn {
                 if cond_dat.ftype != FType::bool {
                     self.permissive_error(
                         line,
-                        logger,
+                        logger.clone(),
                         LogLevel::Error(ErrKind::WhileLoopNotBool(cond_dat.ftype)),
                         LogLevel::Warning(WarnKind::WhileLoopNotBool),
                     );
@@ -464,12 +501,20 @@ impl SemAn {
             }
             AstNode::BreakLoop => {
                 if self.parsing_loop.len() == 0 {
-                    logger.emit(LogLevel::Error(ErrKind::BreakNotLoop), line);
+                    logger.send(LogMessage::new(
+                        LogLevel::Error(ErrKind::BreakNotLoop), 
+                        line,
+                        self.fname.clone()
+                    ));
                 }
             }
             AstNode::ContinueLoop => {
                 if self.parsing_loop.len() == 0 {
-                    logger.emit(LogLevel::Error(ErrKind::ContinueNotLoop), line);
+                    logger.send(LogMessage::new(
+                        LogLevel::Error(ErrKind::ContinueNotLoop), 
+                        line,
+                        self.fname.clone()
+                    ));
                 }
             }
             AstNode::Function {
@@ -481,21 +526,23 @@ impl SemAn {
             } => {
                 let mut override_flag = false;
                 if let Some(v) = &self.parsing_func {
-                    self.parsing_func = Some((name.clone(), v.1));
+                    self.parsing_func = Some((name.path_to_string(), v.1));
                     override_flag = true;
                 } else {
-                    self.parsing_func = Some((name.clone(), 0));
+                    self.parsing_func = Some((name.path_to_string(), 0));
                 }
 
-                if let Some(func_line) = self.declared_parse.get(name) {
+                if let Some(func_line) = self.declared_parse.get(&name.path_to_string()) {
                     if !override_flag {
-                        logger.emit(
-                            LogLevel::Error(ErrKind::FuncRedecl(name.clone(), *func_line)),
+                        logger.send(LogMessage::new(
+                            LogLevel::Error(ErrKind::FuncRedecl(name.path_to_string(),
+                                *func_line)),
                             line,
-                        );
+                            self.fname.clone()
+                        ));
                     }
                 } else {
-                    self.declared_parse.insert(name.clone(), line);
+                    self.declared_parse.insert(name.path_to_string(), line);
                 }
 
                 self.symb_table.enter_scope();
@@ -514,21 +561,23 @@ impl SemAn {
             } => {
                 let mut override_flag = false;
                 if let Some(v) = &self.parsing_func {
-                    self.parsing_func = Some((name.clone(), v.1));
+                    self.parsing_func = Some((name.path_to_string(), v.1));
                     override_flag = true;
                 } else {
-                    self.parsing_func = Some((name.clone(), 0));
+                    self.parsing_func = Some((name.path_to_string(), 0));
                 }
 
-                if let Some(func_line) = self.declared_parse.get(name) {
+                if let Some(func_line) = self.declared_parse.get(&name.path_to_string()) {
                     if !override_flag {
-                        logger.emit(
-                            LogLevel::Error(ErrKind::FuncRedecl(name.clone(), *func_line)),
+                        logger.send(LogMessage::new(
+                            LogLevel::Error(ErrKind::FuncRedecl(name.path_to_string(),
+                                *func_line)),
                             line,
-                        );
+                            self.fname.clone()
+                        ));
                     }
                 } else {
-                    self.declared_parse.insert(name.clone(), line);
+                    self.declared_parse.insert(name.path_to_string(), line);
                 }
 
                 self.symb_table.enter_scope();
@@ -546,13 +595,17 @@ impl SemAn {
                 args,
                 idx,
             } => {
-                let func_dat_vec = match self.func_table.get_func(func_name) {
+                let func_dat_vec = match self.func_table.get_func(
+                    &func_name.path_to_string()) {
                     Some(v) => v.clone(),
                     None => {
-                        logger.emit(
-                            LogLevel::Error(ErrKind::UndeclaredFunc(func_name.clone())),
+                        logger.send(LogMessage::new(
+                            LogLevel::Error(ErrKind::UndeclaredFunc(
+                                func_name.path_to_string()
+                            )),
                             line,
-                        );
+                            self.fname.clone()
+                        ));
                         return exprdat;
                     }
                 };
@@ -583,7 +636,23 @@ impl SemAn {
                         } else {
                             Some(over_idx)
                         };
-
+                       
+                        let funame = func_name.path_to_string();
+                        let ov_module = match funame.rfind("::") {
+                            Some(pos) => funame[..pos].to_owned(),
+                            None => funame.to_owned(), 
+                        };
+                        let same_mod = self.module == ov_module;
+                        
+                        if !overload.public && !same_mod {
+                            // TODO: also iterate overloads till we find public one
+                            logger.send(LogMessage::new(
+                                LogLevel::Error(ErrKind::NotPub(func_name.path_to_string())),
+                                line,
+                                self.fname.clone()
+                            ));
+                            return exprdat;
+                        } 
                         self.matched_overloads.insert(*idx, 
                             (ov_idx_op, overload.ret_type)
                         );
@@ -592,10 +661,11 @@ impl SemAn {
                     }
                 }
 
-                logger.emit(
-                    LogLevel::Error(ErrKind::FuncArgsTypeIncompat(func_name.clone())),
+                logger.send(LogMessage::new(
+                    LogLevel::Error(ErrKind::FuncArgsTypeIncompat(func_name.path_to_string())),
                     line,
-                );
+                    self.fname.clone()
+                ));
             }
             AstNode::ReturnVal { val } => {
                 let retval = self.analyze_expr(&*val, logger);
@@ -604,24 +674,33 @@ impl SemAn {
                     if let Some(fv) = self.func_table.get_func(&curf.0) {
                         let f = &fv[curf.1];
                         if retval.ftype != f.ret_type {
-                            logger.emit(
+                            logger.send(LogMessage::new(
                                 LogLevel::Error(ErrKind::IncompatReturn(
                                     curf.0.to_owned(),
                                     f.ret_type,
                                     retval.ftype,
                                 )),
                                 line,
-                            );
+                                self.fname.clone()
+                            ));
                         }
                     };
                 } else {
-                    logger.emit(LogLevel::Error(ErrKind::ReturnOutOfFunc), line);
+                    logger.send(LogMessage::new(
+                        LogLevel::Error(ErrKind::ReturnOutOfFunc), 
+                        line,
+                        self.fname.clone()
+                    ));
                 };
             }
             AstNode::ExprCast { expr, target_type } => {
                 let expr = self.analyze_expr(&AstRoot::new(*expr.clone(), line), logger);
                 if expr.ftype == *target_type {
-                    logger.emit(LogLevel::Warning(WarnKind::ConvSame(expr.ftype)), line);
+                    logger.send(LogMessage::new(
+                        LogLevel::Warning(WarnKind::ConvSame(expr.ftype)), 
+                        line,
+                        self.fname.clone()
+                    ));
                 }
 
                 exprdat.ftype = *target_type;
@@ -630,10 +709,11 @@ impl SemAn {
                 for (idx, node) in nodes.iter().enumerate() {
                     let ndat = self.analyze_expr(&AstRoot::new(node.clone(), line), logger);
                     if ndat.ftype != *ft {
-                        logger.emit(
+                        logger.send(LogMessage::new(
                             LogLevel::Error(ErrKind::IncompatArrType(*ft, ndat.ftype, idx)),
                             line,
-                        );
+                            self.fname.clone()
+                        ));
                     }
                 }
                 let ft_idx = ftype_to_idx(ft);
@@ -646,10 +726,11 @@ impl SemAn {
                 let array_symb = match self.symb_table.get(arr_name.clone()) {
                     Some(v) => v.clone().1,
                     None => {
-                        logger.emit(
+                        logger.send(LogMessage::new(
                             LogLevel::Error(ErrKind::UndeclaredVar(arr_name.to_owned())),
                             line,
-                        );
+                            self.fname.clone()
+                        ));
                         return exprdat;
                     }
                 };
@@ -662,24 +743,36 @@ impl SemAn {
 
                 let idx_exprdat = self.analyze_expr(&idx, logger);
                 if idx_exprdat.ftype != FType::uint {
-                    logger.emit(
+                    logger.send(LogMessage::new(
                         LogLevel::Error(ErrKind::ArrIdxType(arr_name.clone(), idx_exprdat.ftype)),
                         line,
-                    );
+                        self.fname.clone()
+                    ));
                 }
 
                 let el_type_u = match elem_type {
                     Some(v) => v,
                     None => {
-                        logger.emit(
+                        logger.send(LogMessage::new(
                             LogLevel::Error(ErrKind::Internal("Can't match type".to_owned())),
                             line,
-                        );
+                            self.fname.clone()
+                        ));
                         return exprdat;
                     }
                 };
 
                 exprdat.ftype = el_type_u;
+            }
+            AstNode::Module { name, node } => {
+                let old_mod = self.module.clone();
+                self.module = match old_mod.as_str() {
+                    "main" => name.clone(),
+                    other => format!("{}::{}", other, name),
+                };
+                let exprd = self.analyze_expr(node, &logger.clone());
+                self.module = old_mod;
+
             }
             _ => {}
         }
@@ -690,14 +783,14 @@ impl SemAn {
     fn permissive_error(
         &mut self,
         line: usize,
-        logger: &mut Logger,
+        logger: Sender<LogMessage>,
         lerr: LogLevel,
         lwarn: LogLevel,
     ) {
         if self.permissive {
-            logger.emit(lwarn, line);
+            logger.send(LogMessage::new(lwarn, line, self.fname.clone()));
         } else {
-            logger.emit(lerr, line);
+            logger.send(LogMessage::new(lerr, line, self.fname.clone()));
         }
     }