+++ changelog.md
@@ -3,4 +3,5 @@ v0.7.0:
 - update README.md
 - fixed github ci (once again)
 - `src/main.rs` changes for multifile support
-
+- modules finally works!
+- added `examples/5_modules/..` (+test)